Move AddEntityModal
